==> GroupProducts/FinalProducts/0-CodeRepeatability_TOC.ipynb <==
# Code Repeatability
Best Viewed in Jupyter

### 1.   [Initial Data Carpentry](./1-Data%20Carpentry%20Initial.ipynb)
### 2.   [More Data Carpentry](./2-Data%20Carpentry%20and%20Response%20Grouping.ipynb)
### 3.   Univariate Analysis

### [Demographics](./3-Demographics.ipynb) 
### [Internet Use / Video Games](./3-Internet_Use_Video_Games.ipynb)
### [Traditional / Social Media](./3-Traditional_Media_Social_Media.ipynb)
### [Political Engagement](./3-Political_Engagement.ipynb)
### [Social Views](./3-Social_Views.ipynb)

### 4.   Bivariate Analysis

### [Chi Squared Testing](./4-Chi2Testing.ipynb) 
### [Target Variables](./4-Target%20Variables.ipynb) 

### [Progressivism](./4-Social%20Media%2C%20Internet%20%26%20Video%20Game%20use%20vs%20Progressivism.ipynb) 

### [Voted vs. Demographics](./4-TwoWayTables_Voted_Demographics.ipynb) 
### [Voted vs. Internet Use / Video Games](./4-TwoWayTables_Voted_InternetandVideoGames.ipynb)
### [Voted vs. Traditional / Social Media](./4-TwoWayTables_Voted_TraditionalandSocialMedia.ipynb)
### [Voted vs. Political Engagement](./4-TwoWayTables_Voted_PoliticalEngagement.ipynb)
### [Voted vs. Social Views](./4-TwoWayTables_Voted_SocialViews.ipynb)

### [Political Party vs. Demographics](./4-TwoWayTables_Party_Demographics.ipynb) 
### [Political Party vs. Internet Use / Video Games](./4-TwoWayTables_Party_InternetandVideoGames.ipynb)
### [Political Party vs. Traditional / Social Media](./4-TwoWayTables_Party_TraditionalandSocialMedia.ipynb)
### [Political Party vs. Political Engagement](./4-TwoWayTables_Party_PoliticalEngagement.ipynb)
### [Political Party vs. Social Views](./4-TwoWayTables_Party_SocialViews.ipynb)

### [Engagement](./4-TwoWayTable_Traditional_and_SocialMedia.ipynb)

### 5.   [Final Visualizations](./5-Final_Visualizations.ipynb)
