==> GroupProducts/FinalProducts/1-Data Carpentry Initial.ipynb <==
import pandas as pd
import pwd, os

# Since the casestudy_data folder is under our own pawprint, this will dynamically get your pawprint name so the paths
# referenced below work
pprint = pwd.getpwuid(os.getuid()).pw_name

# We have a CSV file that translates the study's variable names (Q1, Q2, etc) into easier-to-understand terms
# Here we create a mapper dict that holds the translations
cnames = pd.read_csv (f'/home/{pprint}/jupyter/casestudy_data/group_3/ColNames.csv')
varNames = cnames['Var Name'].to_list()
dfNames = cnames['DF Name'].to_list()

colNames = dict(zip (varNames, dfNames))

# Ingest data
df = pd.read_table (f'/home/{pprint}/jupyter/casestudy_data/group_3/37023-0003-Data.tsv', engine='python')

# Rename using mapper dict
df=df.rename (mapper=colNames, axis=1)

# Beginning data carpentry - dropping unused columns

data = df.drop(df.loc[:,'serial':'UKPoliticalNewsSourceSocialMedia'].columns, axis=1) #dropping serial, ID, date of week, etc.
data = data.drop(data.loc[:,'facebookFriendsCurrentWorkSchool':'facebookNewsOthersLessEngaged'].columns, axis=1) #dropping facebook qs about friends
data = data.drop(data.loc[:,'twitterNumFollowed':'twitterFollowedPoliticalNonCelebrities'].columns, axis=1) #dropping twitter qs about followers
data = data.drop(data.loc[:,'schoolTeacherEncouragedDiscussion':'importanceVoting'].columns, axis=1) #dropping upbringing qs and importance qs
data = data.drop(data.loc[:,'polKnowledgeJohnBoehner':'polKnowledgeJuliaGillardRadioHost'].columns, axis=1) #dropping specific political qs
data = data.drop(data.loc[:,'AUSPoliticalParty':'UKPoliticalPartyOther'].columns, axis=1) #dropping AUS and UK political parties
data = data.drop(data.loc[:,'AUSAboriginal':'UKOtherEthnicityOther'].columns, axis=1) #dropping AUS and UK demographic
data = data.drop(data.loc[:,'AUSEducation':'haveDependentChildren'].columns, axis=1) #dropping AUS and UK education and upbringing qs
data = data.drop(data.loc[:,'AUSAreaType':'UKAreaType'].columns, axis=1) #dropping AUS and UK area type
data = data.drop(data.loc[:,'Q64':'COMMENTS'].columns, axis=1) #dropping extra end columns


data.info

#1228 rows

data.head()

import joblib

# for dumping 
joblib.dump([data], 'data.pkl')

# for loading
data = joblib.load('data.pkl')
